Use ComWrappers for IFileSaveDialog I extract from FileOpenDialogWrapper base class FileDialogWrapper to handle just IFileDialog and derive FileSaveDialogWrapper form it. Add missing methods
