+++ src/System.Windows.Forms.Primitives/src/Interop/Interop.IID.cs
@@ -12,6 +12,9 @@ internal partial class Interop
         // D57C7288-D4AD-4768-BE02-9D969532D960
         internal static Guid IFileOpenDialog { get; } = new(0xD57C7288, 0xD4AD, 0x4768, 0xBE, 0x02, 0x9D, 0x96, 0x95, 0x32, 0xD9, 0x60);
 
+        // 84BCCD23-5FDE-4CDB-AEA4-AF64B83D78AB
+        internal static Guid IFileSaveDialog { get; } = new(0x84BCCD23, 0x5FDE, 0x4CDB, 0xAE, 0xA4, 0xAF, 0x64, 0xB8, 0x3D, 0x78, 0xAB);
+
         // 973510DB-7D7F-452B-8975-74A85828D354
         internal static Guid IFileDialogEvents { get; } = new(0x973510DB, 0x7D7F, 0x452B, 0x89, 0x75, 0x74, 0xA8, 0x58, 0x28, 0xD3, 0x54);
