+++ src/System.Windows.Forms/src/System/Windows/Forms/SaveFileDialog.cs
@@ -174,16 +174,18 @@ namespace System.Windows.Forms
         private protected override IFileDialog CreateVistaDialog()
         {
             HRESULT hr = Ole32.CoCreateInstance(
-                ref CLSID.FileSaveDialog,
+                in CLSID.FileSaveDialog,
                 IntPtr.Zero,
                 Ole32.CLSCTX.INPROC_SERVER | Ole32.CLSCTX.LOCAL_SERVER | Ole32.CLSCTX.REMOTE_SERVER,
-                ref NativeMethods.ActiveX.IID_IUnknown,
-                out object obj);
+                in NativeMethods.ActiveX.IID_IUnknown,
+                out IntPtr lpDialogUnknownPtr);
             if (!hr.Succeeded())
             {
                 Marshal.ThrowExceptionForHR((int)hr);
             }
 
+            var obj = WinFormsComWrappers.Instance
+                .GetOrCreateObjectForComInstance(lpDialogUnknownPtr, CreateObjectFlags.None);
             return (IFileSaveDialog)obj;
         }
     }
